+++ sdk/formrecognizer/Azure.AI.FormRecognizer/CHANGELOG.md
@@ -5,6 +5,10 @@
 ### New Features
 - Updated the `FormRecognizerModelFactory` class to support missing model types for mocking.
 
+### Breaking changes
+- Renamed `Id` for `Identity` in all the `StartRecognizeIdDocuments` functionalities. For example, the name of the method is now `StartRecognizeIdentityDocuments`.
+- Renamed the model `ReadingOrder` to `FormReadingOrder`.
+
 ## 3.0.1 (2021-04-09)
 
 ### Key Bug Fixes